+++ src/app/app.component.ts
@@ -1,5 +1,7 @@
-import { AfterViewInit, Component, ElementRef, ViewChild } from '@angular/core';
+import { AfterViewInit, Component, ElementRef, OnInit, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
+import { Store } from '@ngrx/store';
+import { AppState } from './core/@ngrx/app.state';
 import { AppSettingsService } from './core/services/app-settings.service';
 
 @Component({
@@ -7,14 +9,17 @@ import { AppSettingsService } from './core/services/app-settings.service';
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
-export class AppComponent implements AfterViewInit {
+export class AppComponent implements AfterViewInit, OnInit {
   @ViewChild('appTitle')
   appTitle!: ElementRef<HTMLHeadingElement>;
 
-  constructor(router: Router, private setting: AppSettingsService) {
+  constructor(private store: Store<AppState> ,router: Router, private setting: AppSettingsService) {
     const replacer = (key: string, value: any): string =>
       typeof value === 'function' ? value.name : value;
     console.log('Routes: ', JSON.stringify(router.config, replacer, 2));
+  }
+  ngOnInit(): void {
+   
   }
 
   ngAfterViewInit() {