+++ src/app/core/@ngrx/app.state.ts
@@ -1,0 +1,7 @@
+import { RouterState } from "@angular/router";
+import { ProductState } from "./products/products.state";
+
+export interface AppState {
+    products: ProductState;
+    router:RouterState
+   }