+++ src/app/core/@ngrx/root-store.module.ts
@@ -1,0 +1,40 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { StoreModule } from '@ngrx/store';
+import { ProductStoreModule } from './products/product-store.module';
+import { EffectsModule } from '@ngrx/effects';
+import { RouterState, StoreRouterConnectingModule } from '@ngrx/router-store';
+import { routerReducers } from './router/router.reducer';
+import { RouterEffects } from './router/router.effects';
+
+
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule,
+    StoreModule.forRoot(routerReducers, {
+      // All checks will automatically be disabled in production builds
+      runtimeChecks: {
+        strictStateImmutability: true, // default value is true
+        strictActionImmutability: true, // default value is true
+        strictStateSerializability: false, // default value is false
+        strictActionSerializability: false, // default value is false
+        strictActionWithinNgZone: true, // default value is false
+        strictActionTypeUniqueness: true // default value is false
+      }
+    }),
+    StoreRouterConnectingModule.forRoot({
+      stateKey: 'router',
+      routerState: RouterState.Minimal
+    }),
+    ProductStoreModule,
+    EffectsModule.forRoot([RouterEffects]),
+
+
+
+
+
+  ]
+})
+export class RootStoreModule { }