+++ src/app/models/product.model.ts
@@ -1,8 +1,10 @@
-export interface ProductModel {
-  id:number;
-  name: string ;
-  description: string ;
-  price: number ;
-  category :string;
-  isAvailable: boolean;
+export class ProductModel {
+  constructor(
+   public id: number = 0,
+   public name: string= '',
+   public description: string='',
+   public price: number=0,
+   public category: string='',
+   public isAvailable: boolean=false,
+  ) { }
 }